Refactoring: move whitelist commands to whitelist package
