+++ src/main/java/voidpointer/spigot/voidwhitelist/command/whitelist/HelpCommand.java
@@ -1,7 +1,8 @@
-package voidpointer.spigot.voidwhitelist.command;
+package voidpointer.spigot.voidwhitelist.command.whitelist;
 
 import voidpointer.spigot.framework.localemodule.Locale;
 import voidpointer.spigot.framework.localemodule.annotation.AutowiredLocale;
+import voidpointer.spigot.voidwhitelist.command.Command;
 import voidpointer.spigot.voidwhitelist.command.arg.Args;
 
 import static voidpointer.spigot.voidwhitelist.message.WhitelistMessage.*;
