+++ src/main/java/voidpointer/spigot/voidwhitelist/command/whitelist/ReloadCommand.java
@@ -1,4 +1,4 @@
-package voidpointer.spigot.voidwhitelist.command;
+package voidpointer.spigot.voidwhitelist.command.whitelist;
 
 import voidpointer.spigot.framework.di.Autowired;
 import voidpointer.spigot.framework.di.Injector;
@@ -7,6 +7,7 @@ import voidpointer.spigot.framework.localemodule.LocaleLog;
 import voidpointer.spigot.framework.localemodule.annotation.AutowiredLocale;
 import voidpointer.spigot.framework.localemodule.config.TranslatedLocaleFile;
 import voidpointer.spigot.voidwhitelist.VoidWhitelistPlugin;
+import voidpointer.spigot.voidwhitelist.command.Command;
 import voidpointer.spigot.voidwhitelist.command.arg.Args;
 import voidpointer.spigot.voidwhitelist.config.WhitelistConfig;
 import voidpointer.spigot.voidwhitelist.event.EventManager;
