+++ src/main/java/voidpointer/spigot/voidwhitelist/command/whitelist/WhitelistCommand.java
@@ -12,11 +12,13 @@
  *
  *   0. You just DO WHAT THE FUCK YOU WANT TO.
  */
-package voidpointer.spigot.voidwhitelist.command;
+package voidpointer.spigot.voidwhitelist.command.whitelist;
 
 import org.bukkit.command.CommandSender;
 import voidpointer.spigot.framework.localemodule.Locale;
 import voidpointer.spigot.framework.localemodule.annotation.AutowiredLocale;
+import voidpointer.spigot.voidwhitelist.command.Command;
+import voidpointer.spigot.voidwhitelist.command.CommandManager;
 import voidpointer.spigot.voidwhitelist.command.arg.Args;
 
 import java.util.List;
